refactor: organise commands in smaller modules
The single `cmd` module was getting unwieldy and difficult to navigate.

A `workflows` subcommand was created to group all the workflow launching commands.
